Fix build.ps1 is exiting when Chocolatey isn't installed

+++ scripts/build.ps1
@@ -83,10 +83,10 @@ foreach ($rId in $rIds) {
 
 # Build Chocolatey package if a win10-x64 SCD was built
 
-If (Test-Path "$artifactsDir\${appName}_win10-x64\${appName}.exe") {
+If (Get-Command "choco" -ErrorAction SilentlyContinue -and Test-Path "$artifactsDir\${appName}_win10-x64\${appName}.exe") {
     mkdir "$PSScriptRoot\..\chocolatey\tools" -Force
     Remove-Item -Force "$artifactsDir\${appName}.*.nupkg"
     Remove-Item -Force -Recurse "$PSScriptRoot\..\chocolatey\tools\*"
     Copy-Item "$artifactsDir\${appName}_win10-x64" "$PSScriptRoot\..\chocolatey\tools" -Recurse
-    choco pack "$PSScriptRoot\..\chocolatey\hello-netcoreapp.nuspec" -out $artifactsDir
+    choco pack "$PSScriptRoot\..\chocolatey\$appName.nuspec" -out $artifactsDir
 }
